refactor: checker 타입 수정 및 함수 위치 이동
- 도메인에 맞게 이동
