+++ frontend/src/api/community/typeChecker.ts
@@ -2,13 +2,12 @@ import { AxiosError } from 'axios';
 
 import { checkType, hasOwnProperties, isArray, isDateYMD, isNumber, isObject, isString } from '@utils';
 
-import type { CommunityArticle } from '@custom-types';
-
-import { type ApiCommunityArticles } from '@api/community';
+import { type ApiCommunityArticle, type ApiCommunityArticles } from '@api/community';
 import { checkMember } from '@api/member/typeChecker';
 
-type CommunityArticleKeys = keyof CommunityArticle;
-export const checkCommunityArticle = (data: unknown): CommunityArticle => {
+type CommunityArticleKeys = keyof ApiCommunityArticle['get']['responseData'];
+
+export const checkCommunityArticle = (data: unknown): ApiCommunityArticle['get']['responseData'] => {
   if (!isObject(data)) throw new AxiosError(`CommunityArticle does not have correct type: object`);
 
   const keys: Array<CommunityArticleKeys> = ['id', 'author', 'title', 'content', 'createdDate', 'lastModifiedDate'];
