+++ frontend/src/api/links/typeChecker.ts
@@ -1,29 +1,9 @@
 import { AxiosError } from 'axios';
 
-import { checkType, hasOwnProperties, isArray, isBoolean, isDateYMD, isNumber, isObject, isString } from '@utils';
+import { checkType, hasOwnProperties, isArray, isBoolean, isObject } from '@utils';
 
-import type { Link } from '@custom-types';
-
+import { checkLink } from '@api/link/typeChecker';
 import { type ApiLinks } from '@api/links';
-import { checkMember } from '@api/member/typeChecker';
-
-type LinkKeys = keyof Link;
-
-export const checkLink = (data: unknown): Link => {
-  if (!isObject(data)) throw new AxiosError(`Link does not have correct type: object`);
-
-  const keys: Array<LinkKeys> = ['id', 'author', 'linkUrl', 'description', 'createdDate', 'lastModifiedDate'];
-  if (!hasOwnProperties(data, keys)) throw new AxiosError('Link does not have some properties');
-
-  return {
-    id: checkType(data.id, isNumber),
-    author: checkMember(data.author),
-    linkUrl: checkType(data.linkUrl, isString),
-    description: checkType(data.description, isString),
-    createdDate: checkType(data.createdDate, isDateYMD),
-    lastModifiedDate: checkType(data.createdDate, isDateYMD),
-  };
-};
 
 type LinksKeys = keyof ApiLinks['get']['responseData'];
